refactor(Game): turned float into double

+++ src/Game.cpp
@@ -98,7 +98,7 @@ void Game::Setup() {
 }
 
 void Game::CastRays() {
-    float rayAngle = player->rotationAngle - (FOV_ANGLE / 2);
+    double rayAngle = player->rotationAngle - (FOV_ANGLE / 2);
     for (int stripId = 0; stripId < NUM_RAYS; stripId++) {
         rays[stripId].Cast(rayAngle, player, map);
         rayAngle += FOV_ANGLE / NUM_RAYS;
